Move tests/rpc/common.rs to mod.rs
