+++ crates/avalanche-types/tests/rpc/database/batch.rs
@@ -1,6 +1,6 @@
 use std::time::Duration;
 
-use crate::rpc::common::*;
+use super::serve_test_database;
 use avalanche_types::subnet::rpc::{
     database::{
         corruptabledb::Database as CorruptableDb,
